+++ src/lib/api/GraphQLQueries-other-queries.ts
@@ -558,7 +558,8 @@ export const otherQueries = {
                             minAmount
                         }
                         priceAssetPrecision
-                        crossAvailable`
+                        crossAvailable
+                        assetCategory`
                         }
                     }
                 }